Cleanup public API
