+++ tt_metal/api/tt-metalium/device.hpp
@@ -16,7 +16,6 @@
 
 #include <hostdevcommon/common_values.hpp>
 #include <hostdevcommon/kernel_structs.h>  // Not used here, but leaked to programming examples
-#include <tt-metalium/work_executor_types.hpp>
 #include <tt-metalium/data_types.hpp>
 #include <tt-metalium/hal_types.hpp>
 #include <tt-metalium/command_queue_interface.hpp>
