+++ tt_metal/api/tt-metalium/mesh_command_queue.hpp
@@ -28,7 +28,6 @@
 #include <tt-metalium/mesh_trace_id.hpp>
 #include <tt-metalium/distributed_host_buffer.hpp>
 #include <tt-metalium/mesh_workload.hpp>
-#include <tt-metalium/multi_producer_single_consumer_queue.hpp>
 #include <tt-metalium/sub_device_types.hpp>
 #include <tt-metalium/vector_aligned.hpp>
 #include <umd/device/tt_core_coordinates.h>
