+++ tt_metal/api/tt-metalium/mesh_device.hpp
@@ -33,7 +33,6 @@
 #include <tt-metalium/small_vector.hpp>
 #include <tt-metalium/sub_device_types.hpp>
 #include <umd/device/types/arch.h>
-#include <tt-metalium/work_executor_types.hpp>
 
 enum class CoreType;
 namespace tt {
