+++ tt_metal/api/tt-metalium/utils.hpp
@@ -25,11 +25,6 @@ void hash_combine(std::size_t& seed, const T& value) {
     seed ^= hasher(value) + 0x9e3779b9 + (seed << 6) + (seed >> 2);
 }
 
-struct DefinesHash {
-    DefinesHash() {}
-    size_t operator()(const std::map<std::string, std::string>& c_defines) const;
-};
-
 template <typename E, std::enable_if_t<std::is_enum<E>::value, bool> = true>
 auto underlying_type(const E& e) {
     return static_cast<typename std::underlying_type<E>::type>(e);
