+++ ttnn/api/ttnn/tensor/tensor_impl.hpp
@@ -29,6 +29,10 @@ namespace tt_metal {
 
 namespace tensor_impl {
 
+// Empty structs to facilitate Tensor template logic.
+struct bfloat4_b {};
+struct bfloat8_b {};
+
 // -----------------------------------------------------------------------------------------------------------------------------------------------
 // ===============================================================================================================================================
 //                                                              Low Level APIs
